+++ src/ACadSharp/Objects/Evaluations/BlockVisibilityParameter.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using ACadSharp.Attributes;
 using ACadSharp.Entities;
-using CSMath;
 
 namespace ACadSharp.Objects.Evaluations
 {
@@ -10,7 +8,7 @@ namespace ACadSharp.Objects.Evaluations
 	/// Represents a BLOCKVISIBILITYPARAMETER object, in AutoCAD used to
 	/// control the visibility state of entities in a dynamic block.
 	/// </summary>
-	public class BlockVisibilityParameter : Block1PtParameter
+	public partial class BlockVisibilityParameter : Block1PtParameter
 	{
 		/// <inheritdoc/>
 		public override string ObjectName => DxfFileToken.ObjectBlockVisibilityParameter;
@@ -33,66 +31,19 @@ namespace ACadSharp.Objects.Evaluations
 		public List<State> States { get; private set; } = new List<State>();
 
 		/// <summary>
-		/// Gets a title for the dialog to select the subblock that is to be set visible.
+		/// Visibility parameter name.
 		/// </summary>
 		[DxfCodeValue(301)]
 		public string Name { get; set; }
 
 		/// <summary>
-		/// Gets a description presumably for the dialog to select the subblock that is to be set visible.
+		/// Visibility parameter description.
 		/// </summary>
 		[DxfCodeValue(302)]
 		public string Description { get; set; }
 
 		[DxfCodeValue(91)]
 		internal bool Value91 { get; set; }
-
-		/// <summary>
-		/// Represents a named state containing <see cref="Entity"/> objects. <br/>
-		/// The state controls the visibility of the entities assigned to it.
-		/// </summary>
-		public class State : ICloneable
-		{
-			/// <summary>
-			/// Gets the name of the state.
-			/// </summary>
-			[DxfCodeValue(303)]
-			public string Name { get; set; }
-
-			/// <summary>
-			/// Get the list of <see cref="Entity"/> objects in this state.
-			/// </summary>
-			[DxfCodeValue(DxfReferenceType.Count, 94)]
-			[DxfCollectionCodeValue(DxfReferenceType.Handle, 332)]
-			public List<Entity> Entities { get; private set; } = new();
-
-			/// <summary>
-			/// Get the list of <see cref="EvaluationExpression"/> objects.
-			/// </summary>
-			[DxfCodeValue(DxfReferenceType.Count, 95)]
-			[DxfCollectionCodeValue(DxfReferenceType.Handle, 333)]
-			public List<EvaluationExpression> Expressions { get; private set; } = new();
-
-			/// <inheritdoc/>
-			public object Clone()
-			{
-				State clone = (State)MemberwiseClone();
-
-				clone.Entities = new List<Entity>();
-				foreach (var item in this.Entities)
-				{
-					clone.Entities.Add((Entity)item.Clone());
-				}
-
-				clone.Expressions = new();
-				foreach (var item in this.Expressions)
-				{
-					clone.Expressions.Add((EvaluationExpression)item.Clone());
-				}
-
-				return clone;
-			}
-		}
 
 		/// <inheritdoc/>
 		public override CadObject Clone()
